Update logicGame.js
Removed shot checks, I figure since the shot check is done in the shoot() method, all that needs to be done is check for input in the input file and then just call shoot. Also added two arrays to use for collision detection, and an updater to check for currently live in-game objects.